# [FIX] stock: speedup _get_inventory_lines_values