--- a/addons/stock/models/stock_inventory.py
+++ b/addons/stock/models/stock_inventory.py
@@ -6,6 +6,7 @@ from odoo.addons.base.models.ir_model import MODULE_UNINSTALL_FLAG
 from odoo.exceptions import UserError, ValidationError
 from odoo.osv import expression
 from odoo.tools import float_compare, float_is_zero
+from odoo.tools.misc import OrderedSet
 
 
 class Inventory(models.Model):
@@ -302,6 +303,7 @@ class Inventory(models.Model):
         self.ensure_one()
         quants_groups = self._get_quantities()
         vals = []
+        product_ids = OrderedSet()
         for (product_id, location_id, lot_id, package_id, owner_id), quantity in quants_groups.items():
             line_values = {
                 'inventory_id': self.id,
@@ -313,8 +315,11 @@ class Inventory(models.Model):
                 'location_id': location_id,
                 'package_id': package_id
             }
-            line_values['product_uom_id'] = self.env['product.product'].browse(product_id).uom_id.id
+            product_ids.add(product_id)
             vals.append(line_values)
+        product_id_to_product = dict(zip(product_ids, self.env['product.product'].browse(product_ids)))
+        for val in vals:
+            val['product_uom_id'] = product_id_to_product[val['product_id']].product_tmpl_id.uom_id.id
         if self.exhausted:
             vals += self._get_exhausted_inventory_lines_vals({(l['product_id'], l['location_id']) for l in vals})
         return vals