When an inventory has lots of non-zero lines,
i.e. lines with non-exhausted products,
`action_start` is currently quite slow
because of the browse inside the for loop
of `_get_inventory_lines_values`.

Speeds this up by moving
the browse outside of the for loop.

(In recent PR odoo/odoo#79587 I only tested with exhausted inventories so I missed
a bottleneck when calling `action_start` on an inventory without exhausted products 😓)

#### speedup

| Number of lines | Before PR | After PR |
|:----------------:|:-----------:|:--------:|
|  10000 | 1m05s | 9.5s | 




--
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr